+++ ClassroomAPI/DAL/Models/Classroom.cs
@@ -11,6 +11,6 @@ public class Classroom
 	public required string Type { get; set; }
 	public required int Capacity { get; set; }
 	public required int Floor { get; set; }
-	public required byte Number { get; set; }
+	public required int Number { get; set; }
 	public bool IsDeleted { get; set; }
 }
